Backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
requests==2.31.0
httpx==0.26.0
celery==5.3.6
redis==5.0.1
pytesseract==0.3.10
Pillow==10.2.0
scikit-learn==1.4.0
pandas==2.2.0
statsmodels==0.14.1
email-validator==2.1.0.post1
numpy==1.26.3
joeflow
